[Feat] #164 count 필드 int로 통일

+++ src/main/java/org/sopt/bofit/domain/post/dto/response/TrendingPostsResponse.java
@@ -23,13 +23,13 @@ public record TrendingPostsResponse (
     String content,
 
     @Schema(description = "댓글 수 ", example = "8")
-    long commentCount,
+    int commentCount,
 
     @Schema(description = "생성 시간")
     LocalDateTime createdAt,
 
     @Schema(description = "좋아요 수")
-    Long likeCount,
+    int likeCount,
 
     @Schema(description = "사용자 좋아요 여부")
     boolean likedByCurrentUser,
